# quiz/api/openapi/admin.yml
openapi: 3.0.0
info:
  title: Quiz App Admin Panel API
  description: API for managing users, quizzes, and game sessions in the Quiz Application
  version: 1.0.0
  contact:
    name: Quiz App Support
    # email:
servers:
  # - url: https://api.quizapp.com/v1
  #   description: Production server
  # - url: https://staging.quizapp.com/v1
  #   description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Quiz:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the quiz
        userId:
          type: integer
          format: int64
          description: ID of the user who created the quiz
        title:
          type: string
          description: Quiz title
        isPublic:
          type: boolean
          description: Whether the quiz is publicly available
        createdBy:
          type: string
          description: Username of the creator
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
      required:
        - id
        - userId
        - title
        - isPublic
        - createdBy
        - createdAt
        - updatedAt
    Question:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the question
        quizId:
          type: string
          format: uuid
          description: ID of the quiz this question belongs to
        text:
          type: string
          description: Question text
        timeLimit:
          type: integer
          description: Time limit in seconds
        points:
          type: integer
          description: Points for correct answer
        position:
          type: integer
          description: Position in the quiz
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
      required:
        - id
        - quizId
        - text
        - timeLimit
        - points
        - position
    Option:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the option
        questionId:
          type: string
          format: uuid
          description: ID of the question this option belongs to
        text:
          type: string
          description: Option text
        isCorrect:
          type: boolean
          description: Whether this is the correct answer
        position:
          type: integer
          description: Position in the options list
      required:
        - id
        - questionId
        - text
        - isCorrect
        - position
...
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
      required:
        - code
        - message
paths:
...
  /admin/quizzes/{id}:
    get:
      summary: Get quiz by ID
      description: Retrieves a specific quiz by ID
      operationId: getQuizById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Quiz ID
      responses:
        '200':
          description: Quiz details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete quiz
      description: Deletes a quiz
      operationId: deleteQuiz
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Quiz ID
      responses:
        '204':
          description: Quiz deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []